requirements.txt
Keras>=2.2.4
Keras-Applications>=1.0.6
Keras-Preprocessing>=1.0.5
numpy>=1.15.4
Pillow>=5.3.0
scipy>=1.1.0
tensorflow>=1.12.0